apps/backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
sqlalchemy==2.0.41
psycopg2==2.9.10
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0
python-dotenv==1.1.1
